pyproject.toml
[project]
name = "glue-jupyter"
classifiers = [
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Visualization",
    "License :: OSI Approved :: BSD License",
]

description = "Jupyter notebook/lab viewers for glue"
readme = "README.rst"
requires-python = ">=3.10"
dynamic = [ "version",]

authors = [
    {name = "Maarten A. Breddels", email = "maartenbreddels@gmail.com"},
    {name =  "Thomas Robitaille"},
]

dependencies = [
    "echo>=0.14.0",
    "glue-core>=1.25.0",
    "ipympl>=0.3.0",
    "ipyvolume>=0.6.0",
    "ipython_genutils>=0.2",  # missing ipyvolume dependency
    "tornado",  # missing ipyvolume dependency
    "ipywidgets>=7.7.0",
    "ipyvue>=1.2.0,<2",
    "ipyvuetify>=1.2.0,<2",
    "bqplot-image-gl>=1.9.2",
    "bqplot>=0.12.17",
    "bqplot-gl",
    "scikit-image",
]

[project.urls]
Homepage = "https://glueviz.org"

[project.optional-dependencies]
recommended = [
        "glue-vispy-viewers[jupyter]>=1.3.0",
]
docs = [
    "sphinx",
    "sphinx-automodapi",
    "numpydoc",
    "nbconvert>=7.0.0",
    "nbsphinx",
    "sphinx-book-theme",
]

test = [
    "pytest",
    "pytest-cov",
]

nbtest = [
    "astroquery",
    "nbconvert>=6.4.5",
    "nbformat",
    "notebook",
    "pyyaml",
]

visualtest = [
    "playwright==1.41.2",
    "pytest-playwright==0.5.2",
    "pytest-mpl>=0.19.0",
    "solara[pytest]>=1.44.1",
    "starlette<1.0"
]

[build-system]
requires = ["setuptools>=61.2",
            "setuptools_scm"]
build-backend = 'setuptools.build_meta'

[tool.setuptools]
zip-safe = false
include-package-data = false

[tool.setuptools.package-data]
"glue_jupyter.table" = [ "*.vue",]
"glue_jupyter.widgets" = [ "*.vue",]
"glue_jupyter.common.state_widgets" = [ "*.vue",]
glue_jupyter = [ "*.vue",]
"glue_jupyter.common" = [ "*.vue",]
"glue_jupyter.ipyvolume.common" = [ "*.vue",]
"glue_jupyter.ipyvolume.scatter" = [ "*.vue",]
"glue_jupyter.ipyvolume.volume" = [ "*.vue",]
"glue_jupyter.bqplot.tests" = [ "data/*.ipynb",]
"glue_jupyter.ipyvolume.tests" = [ "data/*.ipynb",]
"glue_jupyter.matplotlib.tests" = [ "data/*.ipynb",]
"glue_jupyter.tests" = [ "data/*.ipynb",]
"glue_jupyter.icons" = [ "*.svg",]

[tool.setuptools.packages.find]
namespaces = false

[tool.setuptools_scm]
version_file = "glue_jupyter/_version.py"

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.pytest.ini_options]
addopts = "-m 'not notebook'"
markers = [
    "notebook: marks tests as notebook tests (deselected by default, run with '-m notebook')",
    "mpl_image_compare",
]
filterwarnings = [
    "error:::echo.*:",
    "error::DeprecationWarning",
    "ignore:`BqplotScatterLayerState` is deprecated and will be removed in a future version.*:DeprecationWarning:glue_jupyter.bqplot.*:",
    "ignore:metadata .* was set from the constructor:DeprecationWarning:bqscales.*:",
    "ignore:Passing unrecognized arguments to super:DeprecationWarning:traitlets.*:",
    "ignore:Passing unrecognized arguments to super:DeprecationWarning:ipywidgets.*:",
    "ignore:Image.Image.getdata is deprecated and will be removed in Pillow 14 .2027-10.*:DeprecationWarning:pixelmatch.*:",
    "ignore:websockets.server.WebSocketServerProtocol is deprecated:DeprecationWarning",
    "ignore:remove second argument of ws_handler:DeprecationWarning",
    "ignore:((.|\n)*)Sentinel is not a public part of the traitlets API((.|\n)*)",
    "ignore:.*metadata.*was set from the constructor.*:DeprecationWarning:",
    "ignore:Jupyter is migrating its paths to use standard platformdirs:DeprecationWarning:",
]

[tool.gilesbot]
[tool.gilesbot.circleci_artifacts]
enabled = true

[tool.gilesbot.circleci_artifacts.figure_report]
url = "results/fig_comparison.html"
message = "Click details to see the image test comparisons, for py311-test-visual"
report_on_fail = true

[tool.ruff]
# ruff check: pycodestyle, Pyflakes, McCabe, flake8-bugbear, flake8-simplify
lint.select = ["ALL"]
exclude=[
    "*_parsetab.py",
    "*_lextab.py"
]
lint.ignore = [  # NOTE: non-permanent exclusions should be added to `.ruff.toml` instead.

    # flake8-builtins (A) : shadowing a Python built-in.
    # New ones should be avoided and is up to maintainers to enforce.
    "A00",

    # flake8-commas (COM)
    "COM812",  # TrailingCommaMissing

    # pydocstyle (D)
    # Missing Docstrings
    "D102",  # Missing docstring in public method. Don't check b/c docstring inheritance.
    "D105",  # Missing docstring in magic method. Don't check b/c class docstring.
    # Whitespace Issues
    "D200",  # FitsOnOneLine
    # Docstring Content Issues
    "D410",  # BlankLineAfterSection. Using D412 instead.
    "D400",  # EndsInPeriod.  NOTE: might want to revisit this.

    # pycodestyle (E, W)
    "E741",  # AmbiguousVariableName. Physics variables are often poor code variables

    # flake8-fixme (FIX)
    "FIX002",  # Line contains TODO | notes for improvements are OK iff the code works

    # ISC001 shouldn't be used with ruff format
    # https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "ISC001",

    # pep8-naming (N)
    "N803",  # invalid-argument-name. Physics variables are often poor code variables
    "N806",  # non-lowercase-variable-in-function. Physics variables are often poor code variables

    # flake8-self (SLF)
    "SLF001", # private member access

    # flake8-todos (TD)
    "TD002",  # Missing author in TODO

    # Ruff-specific rules (RUF)
    "RUF001",  # ambiguous-unicode-character-string
    "RUF005",  # unpack-instead-of-concatenating-to-collection-literal -- it's not clearly faster.
    "RUF012",  # Mutable class attributes should be annotated with `typing.ClassVar`
    "RUF021",  # Parenthesize `a and b` expressions when chaining `and` and `or` together
    "RUF022",  # `__all__` is not sorted
    "RUF100",  # Unused blanket `noqa` directive
]

[tool.ruff.lint.extend-per-file-ignores]
# implicit-namespace-package. The examples and configuration files are not a package.
".github/workflows/*.py" = ["INP001"]
"setup.py" = ["INP001"]
"docs/conf.py" = ["INP001"]
# `import` should be at the top-level of a file. Most of the following are conditional imports.
"glue_jupyter/__init__.py" = ["PLC0415"]
"glue_jupyter/app.py" = ["PLC0415"]
# Notebook test imports are inside functions to avoid requiring nbconvert for regular tests.
"glue_jupyter/bqplot/tests/test_bqplot.py" = ["PLC0415"]
"glue_jupyter/ipyvolume/tests/test_ipyvolume.py" = ["PLC0415"]
"glue_jupyter/matplotlib/tests/test_matplotlib.py" = ["PLC0415"]
"glue_jupyter/tests/main_test.py" = ["PLC0415"]

[tool.ruff.lint.flake8-annotations]
ignore-fully-untyped = true
mypy-init-return = true

[tool.ruff.lint.flake8-comprehensions]
allow-dict-calls-with-keyword-arguments = true

[tool.ruff.lint.flake8-type-checking]
exempt-modules = []

[tool.ruff.lint.isort]
known-first-party = ["glue_jupyter", "glue_core"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"
